fix findPluginOrigin in offline mode
