+++ apps/studio/src/store/modules/plugins/index.ts
@@ -8,6 +8,8 @@ import {
   PluginSnapshotsModule,
   PluginSnapshotsState,
 } from "@/store/modules/plugins/PluginSnapshotsModule";
+import { PluginOrigin } from "@/services/plugin";
+import globals from "@/common/globals";
 
 export type PluginsState = {
   snapshots: PluginSnapshotsState;
@@ -19,6 +21,23 @@ export const PluginsModule: Module<{}, RootState> = {
   modules: {
     entries: PluginEntriesModule,
     snapshots: PluginSnapshotsModule,
+  },
+  getters: {
+    findPluginOrigin(_state, _getters, rootState): (id: string) => PluginOrigin {
+      const entries: Record<string, PluginOrigin> = {};
+      for (const entry of rootState.plugins.entries.communityEntries) {
+        entries[entry.id] = "community";
+      }
+      for (const entry of rootState.plugins.entries.officialEntries) {
+        entries[entry.id] = "official";
+      }
+      for (const bundledPlugin of globals.plugins.ensureInstalled) {
+        entries[bundledPlugin.id] = "official";
+      }
+      return (id: string) => {
+        return entries[id] || "unlisted";
+      }
+    },
   },
   actions: {
     async initialize(context) {
